fix(backend): added unchecked files

+++ database/alpha-01.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS users (
   id SERIAL PRIMARY KEY,
   username TEXT NOT NULL UNIQUE,
-  created_at INT NOT NULL,
+  created_at BIGINT NOT NULL,
   password_hash BYTEA NOT NULL,
   salt BYTEA NOT NULL
 );
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS repos (
   name TEXT NOT NULL,
   private BOOLEAN NOT NULL,
   owner_id INT REFERENCES users(id) NOT NULL,
-  created_at INT NOT NULL,
+  created_at BIGINT NOT NULL,
   stars INT NOT NULL
 );
 
@@ -20,6 +20,6 @@ CREATE TABLE IF NOT EXISTS pats (
   name TEXT NOT NULL,
   hash BYTEA NOT NULL,
   owner_id INT REFERENCES users(id) NOT NULL,
-  created_at INT NOT NULL,
-  last_used INT,
+  created_at BIGINT NOT NULL,
+  last_used BIGINT
 );
